Added support for Visual Studio 2017 and D3D12.
The build process now compiles for VS2017 and D3D12. The Render Device
for D3D12 has not yet been implemented yet though.

Also fixed a weird error concerning paths, not sure why it wasn't
caught before.

+++ YTools/YFileUtils/FileStream.h
@@ -55,7 +55,7 @@ class FileStream {
 
       mInFile.seekg(current_pos, std::ios_base::beg);
 
-      data.resize(length);
+      data.resize(static_cast<size_t>(length));
       mInFile.read(reinterpret_cast<char*>(data.data()), length);
       return true;
     }
